MemPostings.Commit() before using them
This changes the way we interact with MemPostings, by batching the
additions into "pending" state, and then actually creating them in the
"Commit" phase.

This is intended to reduce lock contention on very high churn instances,
where lots of Add() calls compete with the reads: this way there will be
only one Add() call that will create all pending to be created postings.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>